+++ CHANGELOG.md
@@ -15,6 +15,11 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 - HyP3 prepare and submit methods now include processing options as named parameters
 - Exceptions raised for HyP3 errors will include the HyP3 API response details
+- `asf_search.get_nearest_neighbors` has been revised:
+  - now limited to Sentinel-1 granules
+  - now raises `ASFSearchError` when the reference granule cannot be found
+  - no longer raises `` for reference granules missing state vector information in CMR
+  - results no longer include `perpendicularBaseline` or `temporalBaseline` fields
 
 ### Fixed
 - `get_authenticated_session` now correctly throws `AuthenticationError` when no `.netrc` file