+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 - `asf_search` module will now raise an `exceptions.ASFSearchError` when it encounters problems and
   will include the Search API response details
-- `HyP3` initialization now accepts a `prompt=True` (default `False`) keyword argument
+- `HyP3.__init__` now accepts a `prompt=True` (default `False`) keyword argument
   which will prompt users for their username or password if not provided
 
 ### Changed