+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `asf_search.get_nearest_neighbors` has been revised:
   - now limited to Sentinel-1 granules
   - now raises `ASFSearchError` when the reference granule cannot be found
-  - no longer raises `` for reference granules missing state vector information in CMR
+  - no longer raises `HTTPError` for reference granules missing state vector information in CMR
   - results no longer include `perpendicularBaseline` or `temporalBaseline` fields
 
 ### Fixed